Componentizing the table layout
