+++ components/current-clip-player/index.js
@@ -1,0 +1,2 @@
+export { default as CurrentClipPlayer } from './CurrentClipPlayer';
+export { default } from './CurrentClipPlayer';
