+++ components/current-panel/index.js
@@ -1,0 +1,2 @@
+export { default as CurrentPanel } from './CurrentPanel';
+export { default } from './CurrentPanel';
