+++ components/current-thread-clip-item/CurrentThreadClipItem.js
@@ -1,10 +1,16 @@
 import { Box, Flex, Stack, useColorMode } from '@chakra-ui/core';
 import Avatar from '../avatar';
 
-import { ThreadClipProgress, ThreadClipAuthorInfo, ThreadClipMeta, ThreadClipMenu, ThreadClipText } from './';
+import {
+	CurrentThreadClipItemProgress,
+	CurrentThreadClipItemAuthorInfo,
+	CurrentThreadClipItemMeta,
+	CurrentThreadClipItemMenu,
+	CurrentThreadClipItemText,
+} from '.';
 import { bgColor } from '../../styles/theme';
 
-export default function ThreadClipSummary(props) {
+export default function CurrentThreadClipItem(props) {
 	const { colorMode } = useColorMode();
 	const { text } = props;
 
@@ -19,7 +25,7 @@ export default function ThreadClipSummary(props) {
 				bg={bgColor[colorMode]}
 				overflow="hidden"
 			>
-				{text ? <ThreadClipText text={text} /> : null}
+				{text ? <CurrentThreadClipItemText text={text} /> : null}
 				<Flex w="100%">
 					<Box w="64px" h="64px" bg="blue.500" rounded="lg" mr="2" flexShrink={0}>
 						<Avatar
@@ -37,12 +43,12 @@ export default function ThreadClipSummary(props) {
 						/>
 					</Box>
 					<Flex flexDir="column" justify="space-between" mr="2" flexGrow={1} overflow="hidden">
-						<ThreadClipAuthorInfo />
-						<ThreadClipProgress percentPlayed={75} lengthInSeconds={45} />
-						<ThreadClipMeta />
+						<CurrentThreadClipItemAuthorInfo />
+						<CurrentThreadClipItemProgress percentPlayed={75} lengthInSeconds={45} />
+						<CurrentThreadClipItemMeta />
 					</Flex>
 					<Flex alignItems="center" justifyContent="center" flexShrink={0} width="40px">
-						<ThreadClipMenu />
+						<CurrentThreadClipItemMenu />
 					</Flex>
 				</Flex>
 			</Flex>
