--- components/current-thread-clip-player/index.js
@@ -1,2 +1,0 @@
-export { default as CurrentThreadClipPlayer } from './CurrentThreadClipPlayer';
-export { default } from './CurrentThreadClipPlayer';
