+++ components/queue-panel/index.js
@@ -1,0 +1,2 @@
+export { default as QueuePanel } from './QueuePanel';
+export { default } from './QueuePanel';
