+++ components/site-content-layout/index.js
@@ -1,0 +1,3 @@
+export { default as SiteContentLayout } from './SiteContentLayout';
+export { default as Breadcrumbs } from './Breadcrumbs';
+export { default } from './SiteContentLayout';
