+++ components/table-view/index.js
@@ -1,3 +1,8 @@
+export { default as TableThreadClipItem } from './TableThreadClipItem';
+export { default as TableThreadClipListHeader } from './TableThreadClipListHeader';
+export { default as TableThreadItemHeader } from './TableThreadItemHeader';
+export { default as TableThreadItem } from './TableThreadItem';
+export { default as TableThreads } from './TableThreads';
+export { default as TableInfo } from './TableInfo';
 export { default as TableView } from './TableView';
-export { default as TableView2 } from './TableView2';
 export { default } from './TableView';
