--- components/thread-clip-summary/index.js
@@ -1,7 +1,0 @@
-export { default as ThreadClipProgress } from './ThreadClipProgress';
-export { default as ThreadClipAuthorInfo } from './ThreadClipAuthorInfo';
-export { default as ThreadClipMeta } from './ThreadClipMeta';
-export { default as ThreadClipMenu } from './ThreadClipMenu';
-export { default as ThreadClipText } from './ThreadClipText';
-export { default as ThreadClipSummary } from './ThreadClipSummary';
-export { default } from './ThreadClipSummary';
